[RB-2594] Refactor rosbag validator
