+++ cartographer_ros/cartographer_ros/rosbag_validate_main.cc
@@ -25,7 +25,9 @@
 #include "cartographer_ros/msg_conversion.h"
 #include "gflags/gflags.h"
 #include "glog/logging.h"
+#include "message_checker.h"
 #include "nav_msgs/Odometry.h"
+#include "range_data_checker.h"
 #include "ros/ros.h"
 #include "ros/time.h"
 #include "rosbag/bag.h"
@@ -39,6 +41,7 @@
 #include "tf2_ros/buffer.h"
 #include "urdf/model.h"
 
+
 DEFINE_string(bag_filename, "", "Bag to process.");
 DEFINE_bool(dump_timing, false,
             "Dump per-sensor timing information in files called "
@@ -47,6 +50,13 @@ DEFINE_bool(dump_timing, false,
 namespace cartographer_ros {
 namespace {
 
+const double kMinAverageAcceleration = 9.5;
+const double kMaxAverageAcceleration = 10.5;
+const double kMaxGapPointsData = 0.1;
+const double kMaxGapImuData = 0.05;
+const double kTimeDeltaSerializationSensorWarning = 0.1;
+const double kTimeDeltaSerializationSensorError = 0.5;
+
 struct FrameProperties {
   ros::Time last_timestamp;
   std::string topic;
@@ -54,22 +64,6 @@ struct FrameProperties {
   std::unique_ptr<std::ofstream> timing_file;
   std::string data_type;
 };
-
-const double kMinLinearAcceleration = 3.;
-const double kMaxLinearAcceleration = 30.;
-const double kTimeDeltaSerializationSensorWarning = 0.1;
-const double kTimeDeltaSerializationSensorError = 0.5;
-const double kMinAverageAcceleration = 9.5;
-const double kMaxAverageAcceleration = 10.5;
-const double kMaxGapPointsData = 0.1;
-const double kMaxGapImuData = 0.05;
-const std::set<std::string> kPointDataTypes = {
-    std::string(
-        ros::message_traits::DataType<sensor_msgs::PointCloud2>::value()),
-    std::string(ros::message_traits::DataType<
-                sensor_msgs::MultiEchoLaserScan>::value()),
-    std::string(
-        ros::message_traits::DataType<sensor_msgs::LaserScan>::value())};
 
 std::unique_ptr<std::ofstream> CreateTimingFile(const std::string& frame_id) {
   auto timing_file = absl::make_unique<std::ofstream>(
@@ -95,160 +89,25 @@ std::unique_ptr<std::ofstream> CreateTimingFile(const std::string& frame_id) {
   return timing_file;
 }
 
-void CheckImuMessage(const sensor_msgs::Imu& imu_message) {
-  auto linear_acceleration = ToEigen(imu_message.linear_acceleration);
-  if (std::isnan(linear_acceleration.norm()) ||
-      linear_acceleration.norm() < kMinLinearAcceleration ||
-      linear_acceleration.norm() > kMaxLinearAcceleration) {
-    LOG_FIRST_N(WARNING, 3)
-        << "frame_id " << imu_message.header.frame_id << " time "
-        << imu_message.header.stamp.toNSec() << ": IMU linear acceleration is "
-        << linear_acceleration.norm() << " m/s^2,"
-        << " expected is [" << kMinLinearAcceleration << ", "
-        << kMaxLinearAcceleration << "] m/s^2."
-        << " (It should include gravity and be given in m/s^2.)"
-        << " linear_acceleration " << linear_acceleration.transpose();
+void logRangeError(const RangeDataChecker::RangeErrorReport& error_report) {
+  if (error_report.error_msg.empty()) {
+    return;
+  }
+  if (error_report.non_sequential || error_report.repeated) {
+    LOG_FIRST_N(ERROR, 3) << error_report.error_msg;
+  } else {
+    LOG_FIRST_N(WARNING, 3) << error_report.error_msg;
   }
 }
 
-bool IsValidPose(const geometry_msgs::Pose& pose) {
-  return ToRigid3d(pose).IsValid();
-}
-
-void CheckOdometryMessage(const nav_msgs::Odometry& message) {
-  const auto& pose = message.pose.pose;
-  if (!IsValidPose(pose)) {
-    LOG_FIRST_N(ERROR, 3) << "frame_id " << message.header.frame_id << " time "
-                          << message.header.stamp.toNSec()
-                          << ": Odometry pose is invalid."
-                          << " pose " << pose;
+void PrintOverlapReport(const RangeDataChecker::FrameToOverlapMap& map) {
+  for (auto& it : map) {
+    LOG(WARNING) << "Sensor with frame_id \"" << it.first
+                  << "\" range measurements have longest overlap of "
+                  << it.second << " s";
   }
 }
 
-void CheckTfMessage(const tf2_msgs::TFMessage& message) {
-  for (const auto& transform : message.transforms) {
-    if (transform.header.frame_id == "map") {
-      LOG_FIRST_N(ERROR, 1)
-          << "Input contains transform message from frame_id \""
-          << transform.header.frame_id << "\" to child_frame_id \""
-          << transform.child_frame_id
-          << "\". This is almost always output published by cartographer and "
-             "should not appear as input. (Unless you have some complex "
-             "remove_frames configuration, this is will not work. Simplest "
-             "solution is to record input without cartographer running.)";
-    }
-  }
-}
-
-bool IsPointDataType(const std::string& data_type) {
-  return (kPointDataTypes.count(data_type) != 0);
-}
-
-class RangeDataChecker {
- public:
-  template <typename MessageType>
-  void CheckMessage(const MessageType& message) {
-    const std::string& frame_id = message.header.frame_id;
-    ros::Time current_time_stamp = message.header.stamp;
-    RangeChecksum current_checksum;
-    cartographer::common::Time time_from, time_to;
-    ReadRangeMessage(message, &current_checksum, &time_from, &time_to);
-    auto previous_time_to_it = frame_id_to_previous_time_to_.find(frame_id);
-    if (previous_time_to_it != frame_id_to_previous_time_to_.end() &&
-        previous_time_to_it->second >= time_from) {
-      if (previous_time_to_it->second >= time_to) {
-        LOG_FIRST_N(ERROR, 3) << "Sensor with frame_id \"" << frame_id
-                              << "\" is not sequential in time."
-                              << "Previous range message ends at time "
-                              << previous_time_to_it->second
-                              << ", current one at time " << time_to;
-      } else {
-        LOG_FIRST_N(WARNING, 3)
-            << "Sensor with frame_id \"" << frame_id
-            << "\" measurements overlap in time. "
-            << "Previous range message, ending at time stamp "
-            << previous_time_to_it->second
-            << ", must finish before current range message, "
-            << "which ranges from " << time_from << " to " << time_to;
-      }
-      double overlap = cartographer::common::ToSeconds(
-          previous_time_to_it->second - time_from);
-      auto it = frame_id_to_max_overlap_duration_.find(frame_id);
-      if (it == frame_id_to_max_overlap_duration_.end() ||
-          overlap > frame_id_to_max_overlap_duration_.at(frame_id)) {
-        frame_id_to_max_overlap_duration_[frame_id] = overlap;
-      }
-    }
-    frame_id_to_previous_time_to_[frame_id] = time_to;
-    if (current_checksum.first == 0) {
-      return;
-    }
-    auto it = frame_id_to_range_checksum_.find(frame_id);
-    if (it != frame_id_to_range_checksum_.end()) {
-      RangeChecksum previous_checksum = it->second;
-      if (previous_checksum == current_checksum) {
-        LOG_FIRST_N(ERROR, 3)
-            << "Sensor with frame_id \"" << frame_id
-            << "\" sends exactly the same range measurements multiple times. "
-            << "Range data at time " << current_time_stamp
-            << " equals preceding data with " << current_checksum.first
-            << " points.";
-      }
-    }
-    frame_id_to_range_checksum_[frame_id] = current_checksum;
-  }
-
-  void PrintReport() {
-    for (auto& it : frame_id_to_max_overlap_duration_) {
-      LOG(WARNING) << "Sensor with frame_id \"" << it.first
-                   << "\" range measurements have longest overlap of "
-                   << it.second << " s";
-    }
-  }
-
- private:
-  typedef std::pair<size_t /* num_points */, Eigen::Vector4f /* points_sum */>
-      RangeChecksum;
-
-  template <typename MessageType>
-  static void ReadRangeMessage(const MessageType& message,
-                               RangeChecksum* range_checksum,
-                               cartographer::common::Time* from,
-                               cartographer::common::Time* to) {
-    auto point_cloud_time = ToPointCloudWithIntensities(message);
-    const cartographer::sensor::TimedPointCloud& point_cloud =
-        std::get<0>(point_cloud_time).points;
-    *to = std::get<1>(point_cloud_time);
-    *from = *to + cartographer::common::FromSeconds(point_cloud[0].time);
-    Eigen::Vector4f points_sum = Eigen::Vector4f::Zero();
-    for (const auto& point : point_cloud) {
-      points_sum.head<3>() += point.position;
-      points_sum[3] += point.time;
-    }
-    if (point_cloud.size() > 0) {
-      double first_point_relative_time = point_cloud[0].time;
-      double last_point_relative_time = (*point_cloud.rbegin()).time;
-      if (first_point_relative_time > 0) {
-        LOG_FIRST_N(ERROR, 1)
-            << "Sensor with frame_id \"" << message.header.frame_id
-            << "\" has range data that has positive relative time "
-            << first_point_relative_time << " s, must negative or zero.";
-      }
-      if (last_point_relative_time != 0) {
-        LOG_FIRST_N(INFO, 1)
-            << "Sensor with frame_id \"" << message.header.frame_id
-            << "\" has range data whose last point has relative time "
-            << last_point_relative_time << " s, should be zero.";
-      }
-    }
-    *range_checksum = {point_cloud.size(), points_sum};
-  }
-
-  std::map<std::string, RangeChecksum> frame_id_to_range_checksum_;
-  std::map<std::string, cartographer::common::Time>
-      frame_id_to_previous_time_to_;
-  std::map<std::string, double> frame_id_to_max_overlap_duration_;
-};
 
 void Run(const std::string& bag_filename, const bool dump_timing) {
   rosbag::Bag bag;
@@ -259,7 +118,9 @@ void Run(const std::string& bag_filename, const bool dump_timing) {
   size_t message_index = 0;
   int num_imu_messages = 0;
   double sum_imu_acceleration = 0.;
+  MessageChecker message_checker;
   RangeDataChecker range_data_checker;
+  RangeDataChecker::RangeErrorReport error_report;
   for (const rosbag::MessageInstance& message : view) {
     ++message_index;
     std::string frame_id;
@@ -268,32 +129,47 @@ void Run(const std::string& bag_filename, const bool dump_timing) {
       auto msg = message.instantiate<sensor_msgs::PointCloud2>();
       time = msg->header.stamp;
       frame_id = msg->header.frame_id;
-      range_data_checker.CheckMessage(*msg);
+      if (!range_data_checker.CheckMessage(*msg, error_report)) {
+        logRangeError(error_report);
+      }
     } else if (message.isType<sensor_msgs::MultiEchoLaserScan>()) {
       auto msg = message.instantiate<sensor_msgs::MultiEchoLaserScan>();
       time = msg->header.stamp;
       frame_id = msg->header.frame_id;
-      range_data_checker.CheckMessage(*msg);
+      if (!range_data_checker.CheckMessage(*msg, error_report)) {
+        logRangeError(error_report);
+      }
     } else if (message.isType<sensor_msgs::LaserScan>()) {
       auto msg = message.instantiate<sensor_msgs::LaserScan>();
       time = msg->header.stamp;
       frame_id = msg->header.frame_id;
-      range_data_checker.CheckMessage(*msg);
+      if (!range_data_checker.CheckMessage(*msg, error_report)) {
+        logRangeError(error_report);
+      }
     } else if (message.isType<sensor_msgs::Imu>()) {
       auto msg = message.instantiate<sensor_msgs::Imu>();
       time = msg->header.stamp;
       frame_id = msg->header.frame_id;
-      CheckImuMessage(*msg);
+      std::string error_msg;
+      if (!message_checker.CheckImuMessage(*msg, error_msg)) {
+        LOG_FIRST_N(WARNING, 3) << error_msg;
+      }
       num_imu_messages++;
       sum_imu_acceleration += ToEigen(msg->linear_acceleration).norm();
     } else if (message.isType<nav_msgs::Odometry>()) {
       auto msg = message.instantiate<nav_msgs::Odometry>();
       time = msg->header.stamp;
       frame_id = msg->header.frame_id;
-      CheckOdometryMessage(*msg);
+      std::string error_msg;
+      if (!message_checker.CheckOdometryMessage(*msg, error_msg)) {
+        LOG_FIRST_N(ERROR, 3) << error_msg;
+      }
     } else if (message.isType<tf2_msgs::TFMessage>()) {
       auto msg = message.instantiate<tf2_msgs::TFMessage>();
-      CheckTfMessage(*msg);
+      std::string error_msg;
+      if (!message_checker.CheckTfMessage(*msg, error_msg)) {
+        LOG_FIRST_N(ERROR, 1) << error_msg;
+      }
       continue;
     } else {
       continue;
@@ -356,7 +232,7 @@ void Run(const std::string& bag_filename, const bool dump_timing) {
   }
   bag.close();
 
-  range_data_checker.PrintReport();
+  PrintOverlapReport(range_data_checker.get_max_overlap_durations());
 
   if (num_imu_messages > 0) {
     double average_imu_acceleration = sum_imu_acceleration / num_imu_messages;
@@ -378,7 +254,7 @@ void Run(const std::string& bag_filename, const bool dump_timing) {
     float max_time_delta =
         *std::max_element(frame_properties.time_deltas.begin(),
                           frame_properties.time_deltas.end());
-    if (IsPointDataType(frame_properties.data_type) &&
+    if (message_checker.IsPointDataType(frame_properties.data_type) &&
         max_time_delta > kMaxGapPointsData) {
       LOG(ERROR) << "Point data (frame_id: \"" << entry_pair.first
                  << "\") has a large gap, largest is " << max_time_delta
